ArealInterpolation

Documentation for ArealInterpolation.

ArealInterpolation.AbstractArealInterpolator

ArealInterpolation.Dasymetric

ArealInterpolation.Direct

ArealInterpolation.Pycnophylactic

ArealInterpolation.decompose_to_geoms_and_values

ArealInterpolation.AbstractArealInterpolator — Type

abstract type AbstractArealInterpolator

The abstract type for all areal interpolation methods.

Interface

All AbstractArealInterpolators must implement the following interface:

interpolate(interpolator::AbstractArealInterpolator, target::GI.AbstractPolygon, sources, values::Vector{Vector}, source_rtree)

This interface is not set in stone and can be changed!

TODOS: - extensive vs intensive variables (currently we act as though variables are intensive) - weight methods (sum vs total) - just pass any arbitrary accumulator

ArealInterpolation.Dasymetric — Type

Dasymetric(mask::Raster)

Dasymetric interpolation uses a mask to weight the influence of each polygon.

Depending on the choice of mask, like land-use data, this can prove to be a more accurate interpolation than the direct or pycnophylactic methods.

ArealInterpolation.Direct — Type

Direct()

A simple direct method for areal interpolation.

Takes the area-weighted mean of all source polygons' features, weighted by their areas of intersection with the target polygon.

This method does not allocate a Raster, but it does perform polygon intersection tests.

ArealInterpolation.Pycnophylactic — Type

Pycnophylactic()

A pycnophylactic method for areal interpolation.

Pycnophylactic interpolation (Tobler, 1979) interpolates the source zone attribute to the target zones in a way that avoids sharp discontinuities between neighbouring target zones. It assumes that no sharp boundaries exist in the distribution of the allocated data, which may not be the case, for example, when target zones are divided by linear features (rivers, railways, roads) or are adjacent to waterbodies.

However, it generates intuitively elegant allocations for many urban case studies with many applications (Kounadi, Ristea, Leitner, & Langford, 2018; Comber, Proctor, & Anthony, 2008).

This description was taken in part from the GIS&T Body of Knowledge.

ArealInterpolation.decompose_to_geoms_and_values — Method

decompose_to_geoms_and_values(sources; features = nothing)

Decompose a table or feature collection into geometries and values. Returns (geometries::Vector{Geometry}, values::NamedTuple{Vector}).

values is a namedtuple of each value column in sources. A value column is something whose eltype satisfies isvaluecol, and is currently Union{Number, Missing}.